6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ long int c;
 long long int d;
 float e;
 
-printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of a char: %zu byte(s)\n", (unsigned)sizeof(a));
 printf("Size of a int: %zu byte(s)\n", (unsigned long)sizeof(b));
 printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(c));
 printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(d));
